Gut PingPong::Match
It's always been more trouble than it's worth and generally a source of
confusion. I still don't like all the ongoing match logic living right
on this model as it's become rather heavy, but I'll live with it until
someone suggests something better.
